+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -48,6 +48,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fix undo not resetting all value when undoing the material affectation in LayerLit material
 - Fix for issue that caused gizmos to render in render textures (case 1174395)
 - Fixed the light emissive mesh not updated when the light was disabled/enabled
+- Fixed light and shadow layer sync when setting the HDAdditionalLightData.lightlayersMask property
 
 ### Changed
 - Color buffer pyramid is not allocated anymore if neither refraction nor distortion are enabled